Minor bug fixes - model data types, removed arguments

+++ scripts/train_protonet.py
@@ -75,7 +75,7 @@ evaluation_taskloader = DataLoader(
 # Model #
 #########
 model = get_few_shot_encoder(num_input_channels)
-model.to(device, dtype=torch.double)
+model.to(device, dtype=torch.float)
 
 ############
 # Training #
